% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contract Obligation by Agency.R
\name{bar_primeob_by_agency}
\alias{bar_primeob_by_agency}
\title{Creating a basic, unscaled bar chart of Prime Contract Obligations by Agency}
\usage{
bar_primeob_by_agency(company_name, FY = 1, n_agencies = 6,
  scale = 1e+06, scale_text = "Millions", FY_range, num_size = 3,
  h = 6, w = 11)
}
\arguments{
\item{company_name}{Name of the company, used in filename and chart title}

\item{FY}{Fiscal year to filter out, commonly the current fiscal year due to incomplete data}

\item{n_agencies}{The number of agencies to include in the view; defaults to 6}

\item{scale}{Dollar scale; default to 1000000}

\item{scale_text}{Dollar scale as text; defaults to Millions}

\item{FY_range}{Fiscal Year range included, for example "FY14-FY17"}

\item{num_size}{Size of heading text for agency names, default to 3.  Note that the fewer
the agencies used, the larger the text should be}

\item{h}{Height of saved chart; defaults to 6}

\item{w}{Width of saved chart; defaults to 11.  Note that more agencies requires a greater width}
}
\value{
A bar chart
}
\description{
Creating a basic, unscaled bar chart of Prime Contract Obligations by Agency
}
\details{
The returned chart will determine whether this function is sufficient or a different,
more scaled chart is needed
}
\examples{
bar_primeob_by_agency(company_name = "Leidos", FY = 2018, n_agencies = 5,
                                scale = 1000000000, scale_text = "Billions", FY_range = "FY14-FY17",
                                num_size = 3, h = 7, w = 13)
}
