% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contract Obligation by Agency.R
\name{bar_primeob_by_agency_choosing}
\alias{bar_primeob_by_agency_choosing}
\title{Creating a bar chart of Prime Contract Obligations by Agency looking at an agency
(or sub-agency) chosen by name to display}
\usage{
bar_primeob_by_agency_choosing(company_name, funding_agency_type1,
  funding_agency_name1, funding_agency_type2 = NULL,
  funding_agency_name2 = NULL, funding_agency_type3 = NULL,
  funding_agency_name3 = NULL, funding_agency_type4 = NULL,
  funding_agency_name4 = NULL, funding_agency_type5 = NULL,
  funding_agency_name5 = NULL, FY = 1, scale = 1000000,
  scale_text = "Millions", grid_division = NULL, FY_range,
  num_size = 3, h = 6, w = 11)
}
\arguments{
\item{company_name}{Name of the company, used in filename and chart title}

\item{funding_agency_type1}{Which field or subfield the agency is listed under}

\item{funding_agency_name1}{The specific agency to display}

\item{funding_agency_type2}{optional second agency type; defaults to null}

\item{funding_agency_name2}{optional second agency name; defaults to null}

\item{funding_agency_type3}{optional third agency type; defaults to null}

\item{funding_agency_name3}{optional third agency type; defaults to null}

\item{funding_agency_type4}{optional fourth agency type; defaults to null}

\item{funding_agency_name4}{optional fourth agency name; defaults to null}

\item{funding_agency_type5}{optional fifth agency type; defaults to null}

\item{funding_agency_name5}{optional fifth agency name; defaults to null}

\item{FY}{Fiscal year to filter out, commonly the current fiscal year due to incomplete data}

\item{scale}{Dollar scale; default to 1000000}

\item{scale_text}{Dollar scale as text; defaults to Millions}

\item{grid_division}{respective sizes of ranges to arrange the final grid.  This generally is
the number of agencies in each plot written as a vector}

\item{num_size}{Size of heading text for agency names, default to 3.  Note that the fewer
the agencies used, the larger the text should be}

\item{h}{Height of saved chart; defaults to 6}

\item{w}{Width of saved chart; defaults to 11.  Note that more agencies requires a greater width}
}
\value{
A bar chart, with two differently scales charts of the data
}
\description{
Creating a bar chart of Prime Contract Obligations by Agency looking at an agency
(or sub-agency) chosen by name to display
}
\details{
Unless already known, use the bar_primeob_by_agency() function to determine the Top Range
and Bottom Range agencies.
}
\examples{
bar_primeob_by_agency_choosing(company_name = "LMI Consulting",
                                         funding_agency_type1 = "Funding Bureau",
                                         funding_agency_name1 = "Customs and Border Protection
                                         (CBP)", FY = 2018, scale = 1000000,
                                         scale_text = "Millions", FY_range = "FY14-FY17",
                                         num_size = 3, h = 6, w = 11)

}
