% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contract Obligation by Agency.R
\name{bar_primeob_by_agency_scaling}
\alias{bar_primeob_by_agency_scaling}
\title{Creating a bar chart of Prime Contract Obligations by Agency that is scaled into a Top
Range and Bottom Range (two scalings)}
\usage{
bar_primeob_by_agency_scaling(company_name, FY = 1, n_agencies = 6,
  scale = 1e+06, scale_text = "Millions", FY_range, top_range,
  bottom_range, optional_third_range = NULL, grid_division,
  num_size = 4, h = 6, w = 11)
}
\arguments{
\item{company_name}{Name of the company, used in filename and chart title}

\item{FY}{Fiscal year to filter out, commonly the current fiscal year due to incomplete data}

\item{n_agencies}{The number of agencies to include in the view; defaults to 6}

\item{scale}{Dollar scale; default to 1000000}

\item{scale_text}{Dollar scale as text; defaults to Millions}

\item{FY_range}{Fiscal Year range included, for example "FY14-FY17"}

\item{top_range}{agencies to be included in the first scaling}

\item{bottom_range}{agencies to be included in the second scaling}

\item{optional_third_range}{if scaling data by three ranges, use this field; otherwise defaults to NULL}

\item{grid_division}{respective sizes of ranges to arrange the final grid.  This generally is
the number of agencies in the top_range and the number in the bottom_range written as a vector.}

\item{num_size}{Size of heading text for agency names, default to 3.  Note that the fewer
the agencies used, the larger the text should be}

\item{h}{Height of saved chart; defaults to 6}

\item{w}{Width of saved chart; defaults to 11.  Note that more agencies requires a greater width}
}
\value{
A bar chart, with two differently scales charts of the data
}
\description{
Creating a bar chart of Prime Contract Obligations by Agency that is scaled into a Top
Range and Bottom Range (two scalings)
}
\details{
Unless already known, use the bar_primeob_by_agency() function to determine the Top Range
and Bottom Range agencies.
}
\examples{
bar_primeob_by_agency _scaling(company_name = "Leidos", FY = 2018,
                                         n_agencies = 5, scale = 1000000000,
                                         scale_text = "Billions", FY_range = "FY14-FY17",
                                         top_range = c(1:2,4), bottom_range = c(3,5),
                                         grid_division = c(3,2), num_size = 3,
                                         h = 7, w = 13)


}
