% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Government-Wide and DPAP Visualizations.R
\name{contract_obs_by_quarter}
\alias{contract_obs_by_quarter}
\title{Quarter comparisons for Contract Obligations - Civilian/Defense Breakout}
\usage{
contract_obs_by_quarter(year, num_size = 3, notyear_prop, FY_range,
  title = paste("Contract Obligations by Quarter: ", FY_range, sep = ""),
  subtitle = NULL, h = 6, w = 11, file_ext = ".jpg")
}
\arguments{
\item{year}{year(s) to be included in the chart, as a vector}

\item{num_size}{Size of number text labels, default to 3; labels for total sum equals num_size + 1}

\item{notyear_prop}{The current year, not to include a proportion}

\item{FY_range}{Range of fiscal years displayed for the title, as text string}

\item{title}{Title of chart, as text string; use \n for multiple lines; defaults to paste("Contract Obligations by Quarter: ", FY_range, sep = "")}

\item{subtitle}{Subtitle of chart, as text string; defaults to NULL}

\item{h}{Height of saved chart; defaults to 6}

\item{w}{Width of saved chart; defaults to 11}

\item{file_ext}{file extension, as text string; .jpg or .pdf; defaults to .jpg}
}
\value{
stacked bar chart
}
\description{
Quarter comparisons for Contract Obligations - Civilian/Defense Breakout
}
\details{
The returned chart will display all contract obligations by quarter for a given range of fiscal years
with a breakout by civilian and defense agencies
}
