% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Functions.R
\name{plot.middle}
\alias{plot.middle}
\title{Create a bar graph from contract obligation data}
\usage{
plot.middle(data.frame, facet_var, num_size)
}
\arguments{
\item{data.frame}{data frame}

\item{facet_var}{The facetted variable for facet_grid()}

\item{num_size}{size of the text that labels the total transaction value in the bar chart; generally 3 or 4}
}
\value{
ggplot
}
\description{
Create a bar graph from contract obligation data
}
\details{
This is for any middle plot in a series of plots (at least three), meant to be used with grid.arrange
}
\examples{


}
