% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Functions.R
\name{plot.one}
\alias{plot.one}
\title{Create a bar graph from contract obligation data for one plot with a single scale}
\usage{
plot.one(data.frame, facet_var, num_size, scale_text = "Millions",
  company_name, FY_range)
}
\arguments{
\item{data.frame}{data frame}

\item{facet_var}{The facetted variable for facet_grid()}

\item{num_size}{size of the text that labels the total transaction value in the bar chart; enerally 3 or 4}

\item{scale_text}{The dollar scale in text form; defaults to "Millions"}

\item{company_name}{The name of the vendor}

\item{FY_range}{Fiscal years displayed in the bar chart as a text string; e.g. "FY14-FY17"}
}
\value{
ggplot
}
\description{
Create a bar graph from contract obligation data for one plot with a single scale
}
\examples{
plot.one(data, facet_var = "funding_agency", 3, scale_text = "Millions", "Leidos", "FY14-FY17")


}
