% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Functions.R
\name{process.data.get.sum}
\alias{process.data.get.sum}
\title{Process contract obligations data from BGov, yielding the total transaction value  by funding agency and fiscal year}
\usage{
process.data.get.sum(data.frame, funding_agency_name, funding_agency_type,
  FY = 1, scale = 1e+06)
}
\arguments{
\item{data.frame}{data frame}

\item{funding_agency_name}{The name of the funding agency or subgroup}

\item{funding_agency_type}{The field or subfield the agency is listed under}

\item{FY}{Fiscal year to filter out, commonly the current fiscal year due to incomplete data}

\item{scale}{Dollar scale; default to 1000000}
}
\value{
data frame with the total transaction value (contract obligations) per fiscal year per funding agency or subfield (all at the same level)
}
\description{
Process contract obligations data from BGov, yielding the total transaction value  by funding agency and fiscal year
}
