% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Government-Wide and DPAP Visualizations.R
\name{total_contract_spending}
\alias{total_contract_spending}
\title{Total Contract Spending, divided into services and products, based on type}
\usage{
total_contract_spending(type, num_size = 4, FY_range, h = 6, w = 11,
  file_ext = ".jpg")
}
\arguments{
\item{type}{text string; choose one of "Government Wide", "DoD", or "Civilian"}

\item{num_size}{Size of number text labels, default to 4}

\item{FY_range}{Range of fiscal years displayed for the title, as text string}

\item{h}{Height of saved chart; defaults to 6}

\item{w}{Width of saved chart; defaults to 11}

\item{file_ext}{file extension, as text string; .jpg or .pdf; defaults to .jpg}
}
\value{
a bar chart
}
\description{
Total Contract Spending, divided into services and products, based on type
}
\details{
Returns a stacked bar chart with DPAP services defined and products, with labels for total services and products
}
